Reorganize files
* Move curve_fit_data and mutli_curve_fit_data to curve_fitting.py
* Move other processing functions to `data_processing.py`
